+++ src/api/index.ts
@@ -1,10 +1,11 @@
-import { IProductsProps } from '@/app/page';
+import { IProductsProps } from '@/app/home/page';
 import { IProductItemProps } from '@/components/ProductCard';
 import axios, { AxiosResponse } from 'axios';
 
 export const api = axios.create({
-  baseURL:
-    `${process.env.NEXT_PUBLIC_API_URL}/api` || 'http://localhost:8080/api',
+  baseURL: process.env.NEXT_PUBLIC_API_URL
+    ? `${process.env.NEXT_PUBLIC_API_URL}/api`
+    : 'http://localhost:8080/api/',
 });
 
 export const getProducts = async (
